Fix warnings

+++ tiger/encryption/xor.c
@@ -1,8 +1,8 @@
 #include "xor.h"
 
-void xor_cryptA(unsigned char str[], unsigned char key, DWORD _size) {
+void xor_cryptA(unsigned char str[], unsigned char key, SIZE_T _size) {
 
-	DWORD i = 0;
+	SIZE_T i = 0;
 	for (i; i < _size; i++) {
 		str[i] = str[i] ^ key;
 	}
